# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Tencent_WeKnora_internal_errors.AppError:
    properties:
      code:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.ErrorCode'
      details: {}
      message:
        type: string
    type: object
  github_com_Tencent_WeKnora_internal_errors.ErrorCode:
    enum:
    - 1000
    - 1001
    - 1002
    - 1003
    - 1004
    - 1005
    - 1006
    - 1007
    - 1008
    - 1009
    - 1010
    - 2000
    - 2001
    - 2002
    - 2003
    - 2004
    - 2100
    - 2101
    - 2102
    - 2103
    type: integer
    x-enum-varnames:
    - ErrBadRequest
    - ErrUnauthorized
    - ErrForbidden
    - ErrNotFound
    - ErrMethodNotAllowed
    - ErrConflict
    - ErrTooManyRequests
    - ErrInternalServer
    - ErrServiceUnavailable
    - ErrTimeout
    - ErrValidation
    - ErrTenantNotFound
    - ErrTenantAlreadyExists
    - ErrTenantInactive
    - ErrTenantNameRequired
    - ErrTenantInvalidStatus
    - ErrAgentMissingThinkingModel
    - ErrAgentMissingAllowedTools
    - ErrAgentInvalidMaxIterations
    - ErrAgentInvalidTemperature
...
  github_com_Tencent_WeKnora_internal_types.MCPAdvancedConfig:
    properties:
      retry_count:
        description: 'Number of retries, default: 3'
        type: integer
      retry_delay:
        description: 'Delay between retries in seconds, default: 1'
        type: integer
      timeout:
        description: 'Timeout in seconds, default: 30'
        type: integer
    type: object
  github_com_Tencent_WeKnora_internal_types.MCPAuthConfig:
    properties:
      api_key:
        type: string
      custom_headers:
        additionalProperties:
          type: string
        type: object
      token:
        type: string
    type: object
  github_com_Tencent_WeKnora_internal_types.MCPEnvVars:
    additionalProperties:
      type: string
    type: object
  github_com_Tencent_WeKnora_internal_types.MCPHeaders:
    additionalProperties:
      type: string
    type: object
  github_com_Tencent_WeKnora_internal_types.MCPService:
    properties:
      advanced_config:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.MCPAdvancedConfig'
      auth_config:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.MCPAuthConfig'
      created_at:
        type: string
      deleted_at:
        $ref: '#/definitions/gorm.DeletedAt'
      description:
        type: string
      enabled:
        type: boolean
      env_vars:
        allOf:
        - $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.MCPEnvVars'
        description: Environment variables for stdio
      headers:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.MCPHeaders'
      id:
        type: string
      name:
        type: string
      stdio_config:
        allOf:
        - $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.MCPStdioConfig'
        description: Required for stdio transport
      tenant_id:
        type: integer
      transport_type:
        $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.MCPTransportType'
      updated_at:
        type: string
      url:
        description: 'Optional: required for SSE/HTTP Streamable'
        type: string
    type: object
  github_com_Tencent_WeKnora_internal_types.MCPStdioConfig:
    properties:
      args:
        description: Command arguments array
        items:
          type: string
        type: array
      command:
        description: 'Command: "uvx" or "npx"'
        type: string
    type: object
  github_com_Tencent_WeKnora_internal_types.MCPTransportType:
    enum:
    - sse
    - http-streamable
    - stdio
    type: string
    x-enum-comments:
      MCPTransportHTTPStreamable: HTTP Streamable
      MCPTransportSSE: Server-Sent Events
      MCPTransportStdio: Stdio (Standard Input/Output)
    x-enum-descriptions:
    - Server-Sent Events
    - HTTP Streamable
    - Stdio (Standard Input/Output)
    x-enum-varnames:
    - MCPTransportSSE
    - MCPTransportHTTPStreamable
    - MCPTransportStdio
...
  gorm.DeletedAt:
    properties:
      time:
        type: string
      valid:
        description: Valid is true if Time is not NULL
        type: boolean
    type: object
...
host: localhost:8080
info:
  contact:
    name: WeKnora Github
    url: https://github.com/Tencent/WeKnora
  description: WeKnora 知识库管理系统 API 文档
  termsOfService: http://swagger.io/terms/
  title: WeKnora API
  version: "1.0"
paths:
...
  /mcp-services:
    get:
      consumes:
      - application/json
      description: 获取当前租户的所有MCP服务
      produces:
      - application/json
      responses:
        "200":
          description: MCP服务列表
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 获取MCP服务列表
      tags:
      - MCP服务
    post:
      consumes:
      - application/json
      description: 创建新的MCP服务配置
      parameters:
      - description: MCP服务配置
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Tencent_WeKnora_internal_types.MCPService'
      produces:
      - application/json
      responses:
        "200":
          description: 创建的MCP服务
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 创建MCP服务
      tags:
      - MCP服务
  /mcp-services/{id}:
    delete:
      consumes:
      - application/json
      description: 删除指定的MCP服务
      parameters:
      - description: MCP服务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 删除成功
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 删除MCP服务
      tags:
      - MCP服务
    get:
      consumes:
      - application/json
      description: 根据ID获取MCP服务详情
      parameters:
      - description: MCP服务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: MCP服务详情
          schema:
            additionalProperties: true
            type: object
        "404":
          description: 服务不存在
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 获取MCP服务详情
      tags:
      - MCP服务
    put:
      consumes:
      - application/json
      description: 更新MCP服务配置
      parameters:
      - description: MCP服务ID
        in: path
        name: id
        required: true
        type: string
      - description: 更新字段
        in: body
        name: request
        required: true
        schema:
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: 更新后的MCP服务
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 更新MCP服务
      tags:
      - MCP服务
  /mcp-services/{id}/resources:
    get:
      consumes:
      - application/json
      description: 获取MCP服务提供的资源列表
      parameters:
      - description: MCP服务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 资源列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 获取MCP服务资源列表
      tags:
      - MCP服务
  /mcp-services/{id}/test:
    post:
      consumes:
      - application/json
      description: 测试MCP服务是否可以正常连接
      parameters:
      - description: MCP服务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 测试结果
          schema:
            additionalProperties: true
            type: object
        "400":
          description: 请求参数错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 测试MCP服务连接
      tags:
      - MCP服务
  /mcp-services/{id}/tools:
    get:
      consumes:
      - application/json
      description: 获取MCP服务提供的工具列表
      parameters:
      - description: MCP服务ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 工具列表
          schema:
            additionalProperties: true
            type: object
        "500":
          description: 服务器错误
          schema:
            $ref: '#/definitions/github_com_Tencent_WeKnora_internal_errors.AppError'
      security:
      - Bearer: []
      summary: 获取MCP服务工具列表
      tags:
      - MCP服务
...
securityDefinitions:
  Bearer:
    description: 输入 Bearer {token} 格式的 JWT 令牌
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"